backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic
pydantic[email]
python-jose[cryptography]
argon2-cffi
python-dotenv
pytest
pytest-asyncio
httpx